// src/compass/privacy/types.ts
export enum AuditAction {
  CREATE = "CREATE",
  READ = "READ",
  UPDATE = "UPDATE",
  DELETE = "DELETE",
  EXPORT = "EXPORT",
  CONSENT_GRANTED = "CONSENT_GRANTED",
  CONSENT_REVOKED = "CONSENT_REVOKED",
  DELETION_REQUESTED = "DELETION_REQUESTED",
  DELETION_COMPLETED = "DELETION_COMPLETED",
}